+++ src/graphql/models.ts
@@ -1,5 +1,5 @@
 export { MylistModel } from "../resolvers/mylist/model.js";
-export { MylistRegistrationModel } from "../resolvers/mylist_registration/models.js";
+export { MylistRegistrationModel } from "../resolvers/MylistRegistration/models.js";
 export { NicovideoVideoSourceModel } from "../resolvers/NicovideoVideoSource/model.js";
 export { TagModel } from "../resolvers/tag/model.js";
 export { UserModel } from "../resolvers/user/model.js";
