+++ src/resolvers/MylistRegistration/MylistRegistration.graphql
