+++ src/resolvers/MylistRegistration/index.ts
