+++ src/resolvers/index.ts
@@ -4,7 +4,7 @@ import { DataSource } from "typeorm";
 import { type Resolvers } from "../graphql/resolvers.js";
 import { resolveMutation } from "./mutation/index.js";
 import { resolveMylist } from "./mylist/index.js";
-import { resolveMylistRegistration } from "./mylist_registration/index.js";
+import { resolveMylistRegistration } from "./MylistRegistration/index.js";
 import { resolveNicovideoVideoSource } from "./NicovideoVideoSource/index.js";
 import { resolveQuery } from "./query/index.js";
 import { resolveTag } from "./tag/index.js";
