+++ src/resolvers/mutation/addVideoToMylist.ts
@@ -4,33 +4,48 @@ import { DataSource } from "typeorm";
 import { ulid } from "ulid";
 
 import { MylistRegistration } from "../../db/entities/mylist_registrations.js";
-import { Mylist } from "../../db/entities/mylists.js";
+import { Mylist, MylistShareRange as MylistEntityShareRange } from "../../db/entities/mylists.js";
 import { Video } from "../../db/entities/videos.js";
 import { MutationResolvers } from "../../graphql/resolvers.js";
 import { addVideoToMylist as addVideoToMylistInNeo4j } from "../../neo4j/add_video_to_mylist.js";
 import { ObjectType, removeIDPrefix } from "../../utils/id.js";
-import { MylistRegistrationModel } from "../mylist_registration/models.js";
+import { MylistRegistrationModel } from "../MylistRegistration/models.js";
+import { MYLIST_NOT_FOUND_OR_PRIVATE_ERROR, MYLIST_NOT_HOLDED_BY_YOU } from "../query/getMylist.js";
 
-export const likeVideo =
-  ({ dataSource, neo4jDriver }: { dataSource: DataSource; neo4jDriver: Neo4jDriver }): MutationResolvers["likeVideo"] =>
-  async (_, { input: { videoId } }, { user }) => {
+export const addVideoToMylist =
+  ({
+    dataSource,
+    neo4jDriver,
+  }: {
+    dataSource: DataSource;
+    neo4jDriver: Neo4jDriver;
+  }): MutationResolvers["addVideoToMylist"] =>
+  async (_parent, { input }, { user }) => {
     if (!user) throw new GraphQLError("need to authenticate");
-
     const mylist = await dataSource
       .getRepository(Mylist)
-      .findOne({ where: { holder: { id: user.id }, isLikeList: true }, relations: { holder: true } });
-    if (!mylist) throw new GraphQLError(`Cannot find favorites list for user for "${user.id}"`);
-
+      .findOne({ where: { id: removeIDPrefix(ObjectType.Mylist, input.mylistId) }, relations: { holder: true } });
+    if (!mylist) {
+      throw new GraphQLError(MYLIST_NOT_FOUND_OR_PRIVATE_ERROR);
+    }
+    if (mylist.holder.id !== user.id) {
+      if (mylist.range === MylistEntityShareRange.PRIVATE) {
+        throw new GraphQLError(MYLIST_NOT_FOUND_OR_PRIVATE_ERROR);
+      } else {
+        throw new GraphQLError(MYLIST_NOT_HOLDED_BY_YOU);
+      }
+    }
     const video = await dataSource
       .getRepository(Video)
-      .findOne({ where: { id: removeIDPrefix(ObjectType.Video, videoId) } });
-    if (!video) throw new GraphQLError(`Cannot find video for "${videoId}"`);
-
+      .findOne({ where: { id: removeIDPrefix(ObjectType.Video, input.videoId) } });
+    if (video === null) {
+      throw new GraphQLError("Video not found");
+    }
     const registration = new MylistRegistration();
     registration.id = ulid();
     registration.mylist = mylist;
     registration.video = video;
-    registration.note = null;
+    registration.note = input.note ?? null;
     await dataSource.getRepository(MylistRegistration).insert(registration);
 
     await addVideoToMylistInNeo4j(neo4jDriver)({
