+++ src/resolvers/mutation/registerTag.test.ts
@@ -10,7 +10,7 @@ import { Tag } from "../../db/entities/tags.js";
 import { User } from "../../db/entities/users.js";
 import { migrations } from "../../db/migrations.js";
 import { TagModel } from "../tag/model.js";
-import { calcNameParentPair, registerTag } from "./register_tag.js";
+import { calcNameParentPair, registerTag } from "./registerTag.js";
 
 describe("Mutation.registerTag", () => {
   test("is defined", () => {
