+++ src/resolvers/mylist/index.ts
@@ -11,7 +11,7 @@ import { Resolvers } from "../../graphql/resolvers.js";
 import { calcMylistIncludeTags } from "../../neo4j/mylist_include_tags.js";
 import { calcRecommendedVideosByMylist } from "../../neo4j/mylist_recommend_videos.js";
 import { addIDPrefix, ObjectType, removeIDPrefix } from "../../utils/id.js";
-import { MylistRegistrationModel } from "../mylist_registration/models.js";
+import { MylistRegistrationModel } from "../MylistRegistration/models.js";
 import { TagModel } from "../tag/model.js";
 import { UserModel } from "../user/model.js";
 import { VideoModel } from "../video/model.js";
