+++ src/resolvers/query/findMylist.ts
